Change qq field to big int

+++ api/models.py
@@ -134,14 +134,14 @@ class RecentGradeCourse(models.Model):
 
 class QQBindInfo(models.Model):
     user = models.ForeignKey(settings.AUTH_USER_MODEL, on_delete=models.CASCADE)
-    qq_id = models.IntegerField(default=0)
+    qq_id = models.BigIntegerField(default=0)
 
     def __str__(self):
         return " ".join([str(self.user), str(self.qq_id)])
 
 
 class NoticeTask(models.Model):
-    qq_id = models.IntegerField(default=0)
+    qq_id = models.BigIntegerField(default=0)
     content = models.TextField()
 
     def __str__(self):
